\name{sleasy}
\alias{sleasy}
\title{Summaries for Lavaan objects made Easy}
\description{
This package takes an lavaan object as input and provides relevant
(change in) fit information in a clear manner and also provides interpretations
of (changes in) model-fit that are easy to understand. The model follows the
fit guidelines by Browne & Cudeck (1993), Chen (2007), Jöreskog and Sörbom
(1993), and Hu & Bentler (1999).
}
\author{
Dr. L.V.D.E. Vogelsmeier & Dr. J. Jongerling

Maintainer: Dr. L.V.D.E. Vogelsmeier <l.v.d.e.vogelsmeier@tilburguniveristy.edu>
& Dr. J. Jongerling <j.jongerling@tilburguniveristy.edu>
}

\references{
  Main literature:

  Browne, M. W., & Cudeck, R. (1993). Alternative ways of assessing model fit. In K. A. Bollen & S. J. Long (Eds.), Testing    structural equation models (pp. 136–162). Sage.

  Chen, F. F. (2007). Sensitivity of goodness of fit indexes to lack of measurement invariance. Structural Equation Modeling:   A Multidisciplinary Journal, 14(3), 464–504. https://doi.org/10.1080/10705510701301834

  Hu, L. t., & Bentler, P. M. (1999). Cutoff criteria for fit indexes in covariance structure analysis: Conventional criteria   versus new alternatives. Structural Equation Modeling: A Multidisciplinary Journal, 6(1), 1–55.
  https://doi.org/10.1080/10705519909540118

  Jöreskog, K. G., & Sörbom, D. (1993). LISREL 8: Structural equation modeling with the SIMPLIS command language. Scientific   Software International.
}
